Fix and simplify the usage of test/where flags
The "value" field happened to work because
most of the primitives like int/string have such field
that returns their own value.

However, in some cases when we would use
a single string to filter tests, it would break
because it would get the individual characters
instead of the actual string, as test filter.